017-Unit-of-Work/internal/usecase/add_course.go
package usecase

import (
	"context"

	"github.com/KelpGF/Go-Expert/017-UOW/internal/entity"
	"github.com/KelpGF/Go-Expert/017-UOW/internal/repository"
	"github.com/google/uuid"
)

type InputUseCase struct {
	CategoryName     string
	CourseName       string
	CourseCategoryID string
}

type AddCourseUseCase struct {
	CourseRepository   repository.CourseRepositoryInterface
	CategoryRepository repository.CategoryRepositoryInterface
}

func NewAddCourseUseCase(courseRepository repository.CourseRepositoryInterface, categoryRepository repository.CategoryRepositoryInterface) *AddCourseUseCase {
	return &AddCourseUseCase{
		CourseRepository:   courseRepository,
		CategoryRepository: categoryRepository,
	}
}

func (a *AddCourseUseCase) Execute(ctx context.Context, input InputUseCase) error {
	category := entity.Category{
		ID:   input.CourseCategoryID,
		Name: input.CategoryName,
	}

	err := a.CategoryRepository.Insert(ctx, category)
	if err != nil {
		return err
	}

	course := entity.Course{
		ID:         uuid.New().String(),
		Name:       input.CourseName,
		CategoryID: input.CourseCategoryID,
	}

	err = a.CourseRepository.Insert(ctx, course)
	if err != nil {
		return err
	}

	return nil
}
